allow compilation for the m68k-atari-mint + m68k-atari-mintelf toolchains

+++ src/libc/termios/tcflush.c
@@ -16,15 +16,14 @@ int tcflush(int fd, int queue_selector) {
   case TCIFLUSH:
   case TCOFLUSH:
   case TCIOFLUSH:
-    selector = (int16_t)queue_selector;
+    selector = queue_selector;
     break;
   default:
     errno = EINVAL;
     return -1;
   }
 
-  int32_t result =
-      Fcntl((int16_t)fd, (int32_t)(intptr_t)&selector, (int16_t)TIOCFLUSH);
+  int32_t result = Fcntl((int16_t)fd, (intptr_t)&selector, (int16_t)TIOCFLUSH);
   if (result < 0) {
     errno = (int)-result;
     return -1;
